+++ pandasdmx/source/__init__.py
@@ -183,7 +183,7 @@ def add_source(info, id=None, override=False, **kwargs):
     info.update(kwargs)
 
     if id in sources and not override:
-        raise ValueError("Data source '%s' already defined; use override=True", id)
+        raise ValueError(f"Data source '{id}' already defined; use override=True")
 
     # Maybe import a subclass that defines a hook
     SourceClass = Source
@@ -207,7 +207,7 @@ def list_sources():
 
 def load_package_sources():
     """Discover all sources listed in ``sources.json``."""
-    with resources.open_binary("pandasdmx", "sources.json") as f:
+    with resources.files("pandasdmx").joinpath("sources.json").open("rb") as f:
         for info in json.load(f):
             add_source(info)